chore: move some buttons from sdk to dogfood app
