+++ packages/react-native-sdk/src/components/Call/CallControls/index.tsx
@@ -6,7 +6,6 @@ export * from './ToggleVideoPreviewButton';
 export * from './ToggleAudioPublishingButton';
 export * from './ToggleVideoPublishingButton';
 export * from './ToggleCameraFaceButton';
-export * from './ChatButton';
 export * from './ReactionsButton';
 export * from './CallControls';
 export * from './CallControlsButton';
@@ -14,6 +13,5 @@ export * from './LobbyControls';
 export * from './IncomingCallControls';
 export * from './OutgoingCallControls';
 export * from './ScreenShareToggleButton';
-export * from './ParticipantsButton';
-export * from './MoreActionsButton';
-export * from './RecordCallButton';
+export * from '../../../../../../sample-apps/react-native/dogfood/src/components/CallControlls/RecordCallButton';
+export * from './BadgeCountIndicator';
