+++ sample-apps/react-native/dogfood/src/components/CallControlls/ChatButton.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
-import { CallControlsButton } from './CallControlsButton';
-import { Chat, IconWrapper } from '../../../icons';
-import { useTheme } from '../../../contexts/ThemeContext';
-import { BadgeCountIndicator } from './BadgeCountIndicator';
+import {
+  CallControlsButton,
+  useTheme,
+} from '@stream-io/video-react-native-sdk';
+import { BadgeCountIndicator } from '@stream-io/video-react-native-sdk/src';
+import { Chat, IconWrapper } from '@stream-io/video-react-native-sdk/src/icons';
 
 /**
  * The props for the Chat Button in the Call Controls.
